moved utility function to handlers folder
