+++ packages/email-notification-dispatcher/src/handlers/agreements/handleAgreementActivatedToProducer.ts
@@ -11,13 +11,13 @@ import {
   eventMailTemplateType,
   getFormattedAgreementStampDate,
   retrieveAgreementDescriptor,
-  retrieveAgreementEservice,
   retrieveHTMLTemplate,
   retrieveTenant,
 } from "../../services/utils.js";
 import {
   AgreementHandlerParams,
   getUserEmailsToNotify,
+  retrieveAgreementEservice,
 } from "../handlerCommons.js";
 
 const notificationType: NotificationType = "agreementManagementToProducer";
